# src/server/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Attr_Article.typeofARTICLE_ATTRS-at-number_": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{}
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"abridged": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					},
					"shortSummary": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{}
						]
					},
					"title": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"subcategory": {
						"type": "string"
					}
				},
				"required": [
					"tags",
					"abridged",
					"summary",
					"shortSummary",
					"title",
					"category",
					"subcategory"
				],
				"type": "object"
			},
			"ArticleAttr": {
				"$ref": "#/components/schemas/Attr_Article.typeofARTICLE_ATTRS-at-number_"
			},
			"DatedAttributes": {
				"properties": {
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object"
			},
			"PostAttributes": {
				"allOf": [
					{
						"$ref": "#/components/schemas/DatedAttributes"
					},
					{
						"properties": {
							"shortSummary": {
								"type": "string"
							},
							"summary": {
								"type": "string"
							},
							"abridged": {
								"type": "string"
							},
							"tags": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"text": {
								"type": "string"
							}
						},
						"required": [
							"shortSummary",
							"summary",
							"abridged",
							"tags",
							"text"
						],
						"type": "object"
					}
				]
			},
			"TitledCategorizedPostAttributes": {
				"allOf": [
					{
						"$ref": "#/components/schemas/PostAttributes"
					},
					{
						"properties": {
							"subcategory": {
								"type": "string"
							},
							"category": {
								"type": "string"
							},
							"title": {
								"type": "string"
							}
						},
						"required": [
							"subcategory",
							"category",
							"title"
						],
						"type": "object"
					}
				]
			},
			"ArticleAttributes": {
				"allOf": [
					{
						"$ref": "#/components/schemas/TitledCategorizedPostAttributes"
					},
					{
						"properties": {},
						"type": "object"
					}
				]
			},
			"ReferralAttributes": {
				"allOf": [
					{
						"$ref": "#/components/schemas/DatedAttributes"
					},
					{
						"properties": {
							"userAgent": {
								"type": "string",
								"description": "the user agent info of the consumer of this referral"
							},
							"target": {
								"type": "string",
								"description": "the url path of the referral destination"
							},
							"referrer": {
								"type": "string",
								"description": "the url path this referral was copied from"
							},
							"referrerId": {
								"type": "number",
								"format": "double",
								"description": "id of a user if ref link was created while logged in"
							}
						},
						"required": [
							"userAgent",
							"target",
							"referrer"
						],
						"type": "object"
					}
				]
			},
			"ReferralCreationAttributes": {
				"$ref": "#/components/schemas/ReferralAttributes"
			},
			"PolicyAttributes": {
				"allOf": [
					{
						"$ref": "#/components/schemas/DatedAttributes"
					},
					{
						"properties": {
							"content": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"content",
							"name"
						],
						"type": "object"
					}
				]
			},
			"Attr_Source.typeofSOURCE_ATTRS-at-number_": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{}
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"abridged": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					},
					"shortSummary": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{}
						]
					},
					"title": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"subcategory": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"alternateTitle": {
						"type": "string"
					}
				},
				"required": [
					"tags",
					"abridged",
					"summary",
					"shortSummary",
					"title",
					"category",
					"subcategory",
					"url",
					"alternateTitle"
				],
				"type": "object"
			},
			"SourceAttr": {
				"$ref": "#/components/schemas/Attr_Source.typeofSOURCE_ATTRS-at-number_"
			},
			"SourceAttributes": {
				"allOf": [
					{
						"$ref": "#/components/schemas/TitledCategorizedPostAttributes"
					},
					{
						"properties": {
							"alternateTitle": {
								"type": "string"
							},
							"rawText": {
								"type": "string"
							},
							"url": {
								"type": "string"
							}
						},
						"required": [
							"alternateTitle",
							"rawText",
							"url"
						],
						"type": "object"
					}
				]
			},
			"SourceReadAndSummarizePayload": {
				"properties": {
					"url": {
						"type": "string"
					}
				},
				"required": [
					"url"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "theskoop-api",
		"version": "0.0.1",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/articles": {
			"get": {
				"operationId": "GetArticles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ArticleAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Articles"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "WriteArticleFromSources",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticleAttributes"
								}
							}
						}
					}
				},
				"tags": [
					"Articles"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/v1/articles/{category}": {
			"get": {
				"operationId": "GetArticlesForCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ArticleAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Articles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "category",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/articles/{category}/{subcategory}": {
			"get": {
				"operationId": "GetArticlesForCategoryAndSubcategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ArticleAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Articles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "category",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "subcategory",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/articles/{category}/{subcategory}/{title}": {
			"get": {
				"operationId": "GetArticleForCategoryAndSubcategoryAndTitle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticleAttributes"
								}
							}
						}
					}
				},
				"tags": [
					"Articles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "category",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "subcategory",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "title",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/referrals": {
			"post": {
				"operationId": "Record",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReferralAttributes"
								}
							}
						}
					}
				},
				"tags": [
					"Referrals"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReferralCreationAttributes"
							}
						}
					}
				}
			}
		},
		"/v1/policies/privacy": {
			"get": {
				"operationId": "GetPrivacyPolicy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PolicyAttributes"
								}
							}
						}
					}
				},
				"tags": [
					"Policies"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/policies/terms": {
			"get": {
				"operationId": "GetTermsOfService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PolicyAttributes"
								}
							}
						}
					}
				},
				"tags": [
					"Policies"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/sources": {
			"get": {
				"operationId": "GetSources",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SourceAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Sources"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "ReadAndSummarizeSource",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SourceAttributes"
								}
							}
						}
					}
				},
				"tags": [
					"Sources"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SourceReadAndSummarizePayload"
							}
						}
					}
				}
			}
		},
		"/v1/sources/{category}": {
			"get": {
				"operationId": "GetSourcesForCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SourceAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Sources"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "category",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/sources/{category}/{subcategory}": {
			"get": {
				"operationId": "GetSourcesForCategoryAndSubCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SourceAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Sources"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "category",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "subcategory",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/sources/{category}/{subcategory}/{title}": {
			"get": {
				"operationId": "GetSourceForCategoryAndSubCategoryAndTitle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SourceAttributes"
								}
							}
						}
					}
				},
				"tags": [
					"Sources"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "category",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "subcategory",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "title",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}